requirements.txt
celery==5.3.6
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
gunicorn==21.2.0
influxdb-client==1.39.0
prometheus-client==0.19.0
psutil==5.9.7
PyJWT==2.10.1
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
SQLAlchemy==2.0.25
streamlit==1.41.1
websockets==12.0
